pyproject.toml
[build-system]
requires = ["uv_build>=0.9.9,<0.10.0"]
build-backend = "uv_build"

[project]
name = "pyloops"
version = "1.8.0"
description = "Unofficial Python SDK for Loops.so email platform"
readme = "README.md"
requires-python = ">=3.12"
license = {text = "MIT"}
authors = [
    {name = "zegerius", email = "aram@askaletta.com"}
]
keywords = ["loops", "email", "api", "sdk", "loops.so"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
]
dependencies = [
    "httpx>=0.23.0,<0.29.0",
    "attrs>=22.2.0",
    "python-dateutil>=2.8.0,<3",
]

[project.optional-dependencies]
testing = ["respx>=0.21.0"]

[project.urls]
Homepage = "https://github.com/doctorgpt-corp/pyloops"
Repository = "https://github.com/doctorgpt-corp/pyloops"
Documentation = "https://loops.so/docs"

[tool.uv.build-backend]
module-root = "src/"

[dependency-groups]
dev = [
    "openapi-python-client>=0.27.1",
    "pyright>=1.1.390",
    "pytest>=8.0",
    "pytest-asyncio>=0.25.0",
    "respx>=0.21.0",
    "ruff>=0.8.0",
]

[tool.ruff]
line-length = 120
target-version = "py312"
src = ["src", "tests"]
exclude = [
    ".venv",
    "src/pyloops/_generated",   # All generated code
]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by formatter)
]

[tool.pyright]
include = ["src"]
exclude = ["**/__pycache__", ".venv"]
typeCheckingMode = "basic"
reportMissingImports = true
reportMissingTypeStubs = false
pythonVersion = "3.12"
